chore: add unit tests for the ranger
