+++ client/packages/grandpa-ranger/jest.setup.js
@@ -1,0 +1,7 @@
+Object.defineProperty(process.env, 'PROFILE', {
+  value: 'local',
+})
+
+Object.defineProperty(process.env, 'CIRCUIT_SIGNER_KEY', {
+  value: 'TEST',
+})
