+++ client/packages/grandpa-ranger/package.json
@@ -9,7 +9,8 @@
     "lint": "eslint --ext .ts src",
     "lint:report": "eslint --ext .ts src --output-file eslint_report.json --format json",
     "start": "ts-node src/index.ts",
-    "watch": "ts-node-dev --respawn  --watch -- src/index.ts"
+    "watch": "ts-node-dev --respawn  --watch -- src/index.ts",
+    "test": "jest"
   },
   "dependencies": {
     "@t3rn/sdk": "^0.2.5",
@@ -21,6 +22,9 @@
     "url": "^0.11.0"
   },
   "devDependencies": {
+    "@types/jest": "^29.5.5",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.1",
     "ts-node-dev": "^2.0.0"
   }
 }
